Split data: status styles
